Refactor and modernize codebase
