+++ src/main/java/pl/net/was/FakerConnectorFactory.java
@@ -41,11 +41,11 @@ public class FakerConnectorFactory
     }
 
     @Override
-    public Connector create(String s, Map<String, String> requiredConfig, ConnectorContext context)
+    public Connector create(String catalogName, Map<String, String> requiredConfig, ConnectorContext context)
     {
         requireNonNull(requiredConfig, "requiredConfig is null");
+        requireNonNull(context, "context is null");
 
-        // A plugin is not required to use Guice; it is just very convenient
         Bootstrap app = new Bootstrap(
                 new FakerModule(
                         context.getNodeManager(),
