+++ src/main/java/pl/net/was/FakerOutputTableHandle.java
@@ -13,48 +13,17 @@
  */
 package pl.net.was;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.trino.spi.connector.ConnectorOutputTableHandle;
 
 import java.util.Set;
 
-import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.util.Objects.requireNonNull;
 
-public final class FakerOutputTableHandle
+public record FakerOutputTableHandle(long table, Set<Long> activeTableIds)
         implements ConnectorOutputTableHandle
 {
-    private final long table;
-    private final Set<Long> activeTableIds;
-
-    @JsonCreator
-    public FakerOutputTableHandle(
-            @JsonProperty("table") long table,
-            @JsonProperty("activeTableIds") Set<Long> activeTableIds)
+    public FakerOutputTableHandle
     {
-        this.table = table;
-        this.activeTableIds = requireNonNull(activeTableIds, "activeTableIds is null");
-    }
-
-    @JsonProperty
-    public long getTable()
-    {
-        return table;
-    }
-
-    @JsonProperty
-    public Set<Long> getActiveTableIds()
-    {
-        return activeTableIds;
-    }
-
-    @Override
-    public String toString()
-    {
-        return toStringHelper(this)
-                .add("table", table)
-                .add("activeTableIds", activeTableIds)
-                .toString();
+        requireNonNull(activeTableIds, "activeTableIds is null");
     }
 }
