+++ src/test/java/pl/net/was/FakerQueryRunner.java
@@ -11,46 +11,60 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package pl.net.was;
 
 import io.airlift.log.Level;
 import io.airlift.log.Logger;
 import io.airlift.log.Logging;
-import io.trino.Session;
 import io.trino.testing.DistributedQueryRunner;
 import io.trino.testing.QueryRunner;
 
 import java.util.Map;
 
+import static io.airlift.testing.Closeables.closeAllSuppress;
 import static io.trino.testing.TestingSession.testSessionBuilder;
 import static java.util.Objects.requireNonNullElse;
 
 public class FakerQueryRunner
 {
+    private static final String CATALOG = "faker";
+
     private FakerQueryRunner() {}
 
-    public static QueryRunner createQueryRunner()
-            throws Exception
+    public static Builder builder()
     {
-        Session defaultSession = testSessionBuilder()
-                .setCatalog("faker")
-                .setSchema("default")
-                .build();
+        return new Builder();
+    }
 
-        QueryRunner queryRunner = DistributedQueryRunner.builder(defaultSession)
-                .setExtraProperties(Map.of(
-                        "http-server.http.port", requireNonNullElse(System.getenv("TRINO_PORT"), "8080")))
-                .setWorkerCount(0)
-                .build();
-        queryRunner.installPlugin(new FakerPlugin());
+    public static class Builder
+            extends DistributedQueryRunner.Builder<Builder>
+    {
+        protected Builder()
+        {
+            super(testSessionBuilder()
+                    .setCatalog(CATALOG)
+                    .setSchema("default")
+                    .build());
+        }
 
-        queryRunner.createCatalog(
-                "faker",
-                "faker",
-                Map.of());
+        @Override
+        public DistributedQueryRunner build()
+                throws Exception
+        {
+            setWorkerCount(0);
+            DistributedQueryRunner queryRunner = super.build();
 
-        return queryRunner;
+            try {
+                queryRunner.installPlugin(new FakerPlugin());
+                queryRunner.createCatalog(CATALOG, "faker");
+
+                return queryRunner;
+            }
+            catch (Exception e) {
+                closeAllSuppress(e, queryRunner);
+                throw e;
+            }
+        }
     }
 
     public static void main(String[] args)
@@ -60,10 +74,13 @@ public class FakerQueryRunner
         logger.setLevel("pl.net.was", Level.DEBUG);
         logger.setLevel("io.trino", Level.INFO);
 
-        QueryRunner queryRunner = createQueryRunner();
+        QueryRunner queryRunner = builder()
+                .setExtraProperties(Map.of(
+                        "http-server.http.port", requireNonNullElse(System.getenv("TRINO_PORT"), "8080")))
+                .build();
 
         Logger log = Logger.get(FakerQueryRunner.class);
         log.info("======== SERVER STARTED ========");
-        log.info("\n====\n%s\n====", ((DistributedQueryRunner) queryRunner).getCoordinator().getBaseUrl());
+        log.info("\n====\n%s\n====", queryRunner.getCoordinator().getBaseUrl());
     }
 }
